requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
frozenlist==1.3.1
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
kombu==5.2.4
multidict==6.0.2
prompt-toolkit==3.0.31
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.8.1
